+++ doc/source/changelog.md
@@ -18,6 +18,7 @@ Primarily developped with Godot 4.4.1+
 - Fixes
     - `VoxelBlockyTypeLibrary`: fixed crash when setting `types` to empty array
     - `VoxelGeneratorGraph`: Editor: fixed error sometimes printing after closing the graph editor
+    - `VoxelStreamRegionFiles`: GDExtension: fixed error creating directories
     - `VoxelTool`: `is_area_editable` was off by one in size, and was always returning `true` if the size of the AABB had any component smaller than 1
     - `VoxelViewer`: reparenting (`remove_child` followed by `add_child`) should no longer reload terrain around the viewer